+++ RELEASES.md
@@ -13,6 +13,7 @@ Upcoming
 [![Build Status](https://travis-ci.org/ibissource/iaf.png)](https://travis-ci.org/ibissource/iaf)
 
 - Fix IFSA no longer gives a warning when the managed connection factory can't be found
+- Add consumes and produces option to rest endpoints to set mediatypes, this also transforms the data from and to JSON/XML when set
 
 7.0-B1
 ---